Added Items + Tweaks

+++ game/entity/Enemy.java
@@ -11,11 +11,12 @@ public class Enemy extends Entity{
 	private Entity target;
 	private double x, y;
 	public int score;
+	protected double targetAngle;
 
 	public Enemy(Color color, Rectangle box, Level level, Entity target) {
 		super(color, box, level);
 		this.target = target;
-		speed = 2;
+		stats.speed = 2;
 		x = box.x;
 		y = box.y;
 	}
@@ -23,20 +24,21 @@ public class Enemy extends Entity{
 	public void kill(){
 		level.entities.remove(this);
 		level.score += 10;
+		level.killed++;
 	}
 	
 	public void onUpdate() {
 		int dx = target.box.x - box.x;
 		int dy = target.box.y - box.y;
 		
-		double angle = Math.toRadians(90) - Math.atan2(dx, dy);
+		targetAngle = Math.toRadians(90) - Math.atan2(dx, dy);
 		
-		double ax = Math.cos(angle);
-		double ay = Math.sin(angle);
+		double ax = Math.cos(targetAngle);
+		double ay = Math.sin(targetAngle);
 		
 		if(!stopped){
-			x += ax * speed;
-			y += ay * speed;			
+			x += ax * stats.speed;
+			y += ay * stats.speed;			
 		}
 		
 		box.x = (int) x;
@@ -46,8 +48,15 @@ public class Enemy extends Entity{
 	public void isTouching(Entity e){
 		if(Projectile.class.isAssignableFrom(e.getClass())){
 			Projectile p = (Projectile) e;
-			damage(p.damage);
-			p.kill();
+			if(p.owner.getClass().equals(Player.class)){
+				if(p.damage > stats.health){
+					p.damage -= stats.health;
+					kill();
+				} else {					
+					damage(p.damage);
+					p.kill();					
+				}
+			}
 		}
 	}
